requirements.txt
# Django admin theme
django-material-admin==1.8.5

# DJANGO REST FRAMEWORK---------------------------------------------------
djangorestframework==3.12.4
django-cors-headers==3.7.0
djangorestframework-api-key==2.0.0
djangorestframework-simplejwt==4.7.1
drf-yasg==1.20.0
drf-rw-serializers==1.0.4

jsonfield==3.1.0
pytest==6.2.4